Refactor of file structure to de-clutter the main src folder.
